separate pipeline and bin callback specifications
