+++ lib/membrane/core/bin/callback_context.ex
@@ -1,4 +1,6 @@
 defmodule Membrane.Core.Bin.CallbackContext do
+  @moduledoc false
+
   use Membrane.Core.CallbackContext,
     playback_state: Membrane.PlaybackState.t(),
     clock: Membrane.Clock.t() | nil,
