+++ lib/membrane/pipeline/callback_context/notification.ex
@@ -1,7 +1,7 @@
 defmodule Membrane.Pipeline.CallbackContext.Notification do
   @moduledoc """
   Structure representing a context that is passed to the callback when
-  bin receives notification.
+  pipeline receives notification.
   """
   use Membrane.Core.Pipeline.CallbackContext
 end
